bbheader_source: use broadcast labels
This modifies the bbheader_source to use broadcast labels (i.e., no
label included in the GSE header) instead of 6-byte labels (or label
re-use when possible).

For the time being, this is a quick an simple modification that
forces this behavior, but this change should really be done with
a configuration parameter in the block constructor.