bbheader_source: add support for reading from a TUN
This adds an option to select between reading packets from a TAP
device with PCAP and reading packets from a TUN device with read().
Other possible options, such as reading packets from a TAP device
with read() have not been implemented, but they could be added
following the same scheme.